pyproject.toml
[tool.poetry]
name = "cat-facts-api-tests"
version = "0.1.0"
description = "Testing framework for the Cat Facts API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cat_facts_api"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pytest = "^7.4.0"
jsonschema = "^4.19.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-html = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
